[feat] 新しいPostDetailおよびPostsコンポーネントを追加し、データ表示を改善
