--- apps/fe/src/app/articles/mockData.ts
@@ -1,76 +1,0 @@
-import type { DBPost } from "./type";
-
-export const mockDBPosts: DBPost[] = [
-  {
-    id: 1,
-    name: "taro",
-    post: "こんにちは、世界！",
-    postTime: "2025-07-23T12:30:00.000Z",
-  },
-  {
-    id: 2,
-    name: "hanako",
-    post: "TypeScript最高！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 3,
-    name: "tozeki",
-    post: "Next.js最高！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 4,
-    name: "tterao",
-    post: "宣言的UI最高！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 5,
-    name: "rikeda",
-    post: "フロントエンド最高！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 6,
-    name: "tmp太郎",
-    post: "aaaaaaaaaaaaa！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 7,
-    name: "tmp太郎",
-    post: "aaaaaaaaaaaaa！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 8,
-    name: "tmp太郎",
-    post: "aaaaaaaaaaaaa！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 9,
-    name: "tmp太郎",
-    post: "aaaaaaaaaaaaa！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 10,
-    name: "tmp太郎",
-    post: "aaaaaaaaaaaaa！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 11,
-    name: "tmp太郎",
-    post: "aaaaaaaaaaaaa！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-  {
-    id: 12,
-    name: "tmp太郎",
-    post: "aaaaaaaaaaaaa！",
-    postTime: "2025-07-23T12:00:50.000Z",
-  },
-];
