--- apps/fe/src/app/articles/type.ts
@@ -1,8 +1,0 @@
-export type Post = {
-  id: number;
-  name: string;
-  post: string;
-  postTime: Date;
-};
-
-export type DBPost = Omit<Post, "postTime"> & { postTime: string };
