+++ apps/fe/src/app/posts/type.ts
@@ -1,0 +1,6 @@
+export type Post = {
+  id: string;
+  name: string;
+  post: string;
+  createdAt: string;
+};
